requirements.txt
pandas==2.2.2
dash_bootstrap_components==1.6.0
dash==2.17.0
MLB-StatsAPI
dash-gif-component==1.1.0